Match shared ownership count type
Expose the shared pointer use count with its standard signed return type. This keeps the native lifetime assertion warning-free across toolchains.

+++ src/cvm/string_intptr_storage.h
@@ -22,7 +22,7 @@ class DabStringIntPtrStorage
         return storage.get();
     }
 
-    size_t use_count() const
+    long use_count() const
     {
         return storage.use_count();
     }
